Replace everything

+++ rclpy/car/bramy/PostProcess.py
@@ -89,7 +89,7 @@ class PostProcessing(Node):
                     self.TrackerPos = x1, x2, y1, y2
             else:
                 multi = True
-                outputs = ocSort.update(bboxes.astype(np.float32), (640,480), (640,480))
+                outputs = ocSort.update(bboxes, (640,480), (640,480))
                 self.get_logger().info(f"{outputs}")
                 self.getCenterBox(outputs)
